flare/requirements.txt
torch
torchvision
hydra-core
omegaconf
rich
moviepy==1.0.3
imageio
ffmpeg
huggingface-hub[hf-transfer,cli]
gymnasium
flask
datasets==3.3.2
gdown
einops
diffusers
wandb
av
pymunk
zarr
robomimic
mujoco==3.3.0
dm_control==1.0.28
fvcore
